feat: optimize tile preloading bounds, distance sorting and concurrent priority
This implements real-time tile preload management:
1. Calculates target tile bounds across z, z+1, and z-1 ensuring exact alignment with the viewport's projected 1x1 screen footprint.
2. Continually scores candidate tiles by scaling their Euclidean distance from the viewport center to normalized values across all zoom levels.
3. Removes 'movestart' and 'zoomstart' aborts in favor of a throttled ongoing execution during 'move' and 'moveend'.
4. Implements a max 5 concurrency logic that continuously preempts (aborts) in-flight preloads if a closer, higher-priority tile enters the viewport during map movement.
