+++ patch.js
@@ -1,0 +1,5 @@
+const fs = require('fs');
+
+let content = fs.readFileSync('src/utils/CachedTileLayer.js', 'utf8');
+
+// I will re-implement the preload logic according to the new requirements
